requirements.txt
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.0
drf-yasg==1.21.3
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.25.1
django-filter==22.1